+++ pbj/implicit_solvent/pb_formulation/formulations/alpha_beta_external_potential.py
