+++ pbj/implicit_solvent/pb_formulation/formulations/direct_stern.py
@@ -52,7 +52,7 @@ def create_stern_mesh(self):
     """
 
     if hasattr(self, "mesh_density"):
-        stern_solute_object = pbj.electrostatics.Solute(
+        stern_solute_object = pbj.implicit_solvent.Solute(
             stern_pqr_file,
             external_mesh_file=None,
             save_mesh_build_files=self.save_mesh_build_files,
@@ -67,7 +67,7 @@ def create_stern_mesh(self):
         )
 
     else:
-        stern_solute_object = pbj.electrostatics.Solute(
+        stern_solute_object = pbj.implicit_solvent.Solute(
             stern_pqr_file,
             external_mesh_file=None,
             save_mesh_build_files=self.save_mesh_build_files,
@@ -275,7 +275,7 @@ def rhs(self):
 def block_diagonal_preconditioner(solute):
     from scipy.sparse import diags, bmat, block_diag, dok_matrix
     from scipy.sparse.linalg import aslinearoperator
-    from pbj.electrostatics.utils import matrix_to_discrete_form, rhs_to_discrete_form
+    from pbj.implicit_solvent.utils import matrix_to_discrete_form, rhs_to_discrete_form
 
     dirichl_space_diel = solute.dirichl_space
     neumann_space_diel = solute.neumann_space
@@ -455,7 +455,7 @@ def block_diagonal_preconditioner(solute):
 
 
 def mass_matrix_preconditioner(solute):
-    from pbj.electrostatics.utils import matrix_to_discrete_form, rhs_to_discrete_form
+    from pbj.implicit_solvent.utils import matrix_to_discrete_form, rhs_to_discrete_form
 
     solute.matrices["A_final"] = solute.matrices["A"]
     solute.rhs["rhs_final"] = [
