+++ pbj/implicit_solvent/pb_formulation/formulations/muller_external.py
@@ -4,7 +4,7 @@ from bempp.api.operators.boundary import sparse, laplace, modified_helmholtz
 from .common import calculate_potential_one_surface
 
 
-invert_potential = False
+invert_potential = True
 
 
 def verify_parameters(self):
@@ -51,10 +51,10 @@ def lhs(self):
     ep = ep_ex / ep_in
 
     A = bempp.api.BlockedOperator(2, 2)
-    A[0, 0] = phi_identity + dlp_in - dlp_ex
-    A[0, 1] = -slp_in + ((1.0 / ep) * slp_ex)
-    A[1, 0] = -hlp_in + (ep * hlp_ex)
-    A[1, 1] = dph_identity - adlp_in + adlp_ex
+    A[0, 0] = phi_identity - dlp_ex + dlp_in
+    A[0, 1] = slp_ex - (ep * slp_in)
+    A[1, 0] = hlp_ex + ((1.0 / ep) * hlp_in)
+    A[1, 1] = dph_identity + adlp_ex - adlp_in
 
     self.matrices["A"] = A
 
@@ -64,6 +64,7 @@ def rhs(self):
     q = self.q
     x_q = self.x_q
     ep_in = self.ep_in
+    ep_ex = self.ep_ex
 
     @bempp.api.real_callable
     def d_green_func(x, n, domain_index, result):
@@ -71,7 +72,9 @@ def rhs(self):
             (x[0] - x_q[:, 0]) ** 2 + (x[1] - x_q[:, 1]) ** 2 + (x[2] - x_q[:, 2]) ** 2
         )
         const = -1.0 / (4.0 * np.pi * ep_in)
-        result[:] = const * np.sum(q * np.dot(x - x_q, n) / (nrm**3))
+        result[:] = (
+            (ep_in / ep_ex) * const * np.sum(q * np.dot(x - x_q, n) / (nrm**3))
+        )
 
     @bempp.api.real_callable
     def green_func(x, n, domain_index, result):
@@ -83,17 +86,17 @@ def rhs(self):
     rhs_1 = bempp.api.GridFunction(dirichl_space, fun=green_func)
     rhs_2 = bempp.api.GridFunction(dirichl_space, fun=d_green_func)
 
-    self.rhs["rhs_1"] = rhs_1
-    self.rhs["rhs_2"] = rhs_2
+    self.rhs["rhs_1"], self.rhs["rhs_2"] = rhs_1, rhs_2
 
 
 def mass_matrix_preconditioner(solute):
-    from pbj.electrostatics.utils import matrix_to_discrete_form, rhs_to_discrete_form
+    from pbj.implicit_solvent.utils import matrix_to_discrete_form, rhs_to_discrete_form
 
-    # Option A:
     """
+    #Option A:
     from bempp.api.utils.helpers import get_inverse_mass_matrix
     from bempp.api.assembly.blocked_operator import BlockedDiscreteOperator
+
     matrix = solute.matrices["A"]
     nrows = len(matrix.range_spaces)
     range_ops = np.empty((nrows, nrows), dtype="O")
