+++ pbj/implicit_solvent/pb_formulation/formulations/muller_internal.py
@@ -1,11 +1,10 @@
 import numpy as np
 import bempp.api
-import os
 from bempp.api.operators.boundary import sparse, laplace, modified_helmholtz
 from .common import calculate_potential_one_surface
 
 
-invert_potential = True
+invert_potential = False
 
 
 def verify_parameters(self):
@@ -52,87 +51,49 @@ def lhs(self):
     ep = ep_ex / ep_in
 
     A = bempp.api.BlockedOperator(2, 2)
-    A[0, 0] = (0.5 * (1 + (1.0 / ep))) * phi_identity - dlp_ex + (1.0 / ep) * dlp_in
-    A[0, 1] = slp_ex - slp_in
-    A[1, 0] = (1.0 / ep) * hlp_ex - (1.0 / ep) * hlp_in
-    A[1, 1] = (0.5 * (1 + (1.0 / ep))) * dph_identity + (1.0 / ep) * adlp_ex - adlp_in
+    A[0, 0] = phi_identity + dlp_in - dlp_ex
+    A[0, 1] = -slp_in + ((1.0 / ep) * slp_ex)
+    A[1, 0] = -hlp_in + (ep * hlp_ex)
+    A[1, 1] = dph_identity - adlp_in + adlp_ex
 
     self.matrices["A"] = A
 
 
 def rhs(self):
     dirichl_space = self.dirichl_space
-    neumann_space = self.neumann_space
     q = self.q
     x_q = self.x_q
-    ep_ex = self.ep_ex
-    rhs_constructor = self.rhs_constructor
+    ep_in = self.ep_in
 
-    if rhs_constructor == "fmm":
+    @bempp.api.real_callable
+    def d_green_func(x, n, domain_index, result):
+        nrm = np.sqrt(
+            (x[0] - x_q[:, 0]) ** 2 + (x[1] - x_q[:, 1]) ** 2 + (x[2] - x_q[:, 2]) ** 2
+        )
+        const = -1.0 / (4.0 * np.pi * ep_in)
+        result[:] = const * np.sum(q * np.dot(x - x_q, n) / (nrm**3))
 
-        @bempp.api.callable(vectorized=True)
-        def rhs1_fun(x, n, domain_index, result):
-            import exafmm.laplace as _laplace
+    @bempp.api.real_callable
+    def green_func(x, n, domain_index, result):
+        nrm = np.sqrt(
+            (x[0] - x_q[:, 0]) ** 2 + (x[1] - x_q[:, 1]) ** 2 + (x[2] - x_q[:, 2]) ** 2
+        )
+        result[:] = np.sum(q / nrm) / (4.0 * np.pi * ep_in)
 
-            sources = _laplace.init_sources(x_q, q)
-            targets = _laplace.init_targets(x.T)
-            fmm = _laplace.LaplaceFmm(p=10, ncrit=500, filename=".rhs.tmp")
-            tree = _laplace.setup(sources, targets, fmm)
-            values = _laplace.evaluate(tree, fmm)
-            os.remove(".rhs.tmp")
-            result[:] = values[:, 0] / ep_ex
+    rhs_1 = bempp.api.GridFunction(dirichl_space, fun=green_func)
+    rhs_2 = bempp.api.GridFunction(dirichl_space, fun=d_green_func)
 
-        @bempp.api.callable(vectorized=True)
-        def rhs2_fun(x, n, domain_index, result):
-            import exafmm.laplace as _laplace
-
-            sources = _laplace.init_sources(x_q, q)
-            targets = _laplace.init_targets(x.T)
-            fmm = _laplace.LaplaceFmm(p=10, ncrit=500, filename=".rhs.tmp")
-            tree = _laplace.setup(sources, targets, fmm)
-            values = _laplace.evaluate(tree, fmm)
-            os.remove(".rhs.tmp")
-            result[:] = np.sum(values[:, 1:] * n.T, axis=1) / ep_ex
-
-        rhs_1 = bempp.api.GridFunction(dirichl_space, fun=rhs1_fun)
-        rhs_2 = bempp.api.GridFunction(neumann_space, fun=rhs2_fun)
-
-    else:
-
-        @bempp.api.real_callable
-        def d_green_func(x, n, domain_index, result):
-            nrm = np.sqrt(
-                (x[0] - x_q[:, 0]) ** 2
-                + (x[1] - x_q[:, 1]) ** 2
-                + (x[2] - x_q[:, 2]) ** 2
-            )
-            const = -1.0 / (4.0 * np.pi * ep_ex)
-            result[:] = const * np.sum(q * np.dot(x - x_q, n) / (nrm**3))
-
-        @bempp.api.real_callable
-        def green_func(x, n, domain_index, result):
-            nrm = np.sqrt(
-                (x[0] - x_q[:, 0]) ** 2
-                + (x[1] - x_q[:, 1]) ** 2
-                + (x[2] - x_q[:, 2]) ** 2
-            )
-            result[:] = np.sum(q / nrm) / (4.0 * np.pi * ep_ex)
-
-        rhs_1 = bempp.api.GridFunction(dirichl_space, fun=green_func)
-        rhs_2 = bempp.api.GridFunction(dirichl_space, fun=d_green_func)
-
-    self.rhs["rhs_1"], self.rhs["rhs_2"] = rhs_1, rhs_2
+    self.rhs["rhs_1"] = rhs_1
+    self.rhs["rhs_2"] = rhs_2
 
 
 def mass_matrix_preconditioner(solute):
-    from pbj.electrostatics.utils import matrix_to_discrete_form, rhs_to_discrete_form
+    from pbj.implicit_solvent.utils import matrix_to_discrete_form, rhs_to_discrete_form
 
     # Option A:
     """
     from bempp.api.utils.helpers import get_inverse_mass_matrix
     from bempp.api.assembly.blocked_operator import BlockedDiscreteOperator
-
-
     matrix = solute.matrices["A"]
     nrows = len(matrix.range_spaces)
     range_ops = np.empty((nrows, nrows), dtype="O")
