+++ pbj/implicit_solvent/pb_formulation/formulations/slic_prop.py
@@ -9,21 +9,21 @@ def verify_parameters(self):
 
 
 def lhs(self):
-    pbj.electrostatics.pb_formulation.formulations.direct_stern.lhs(self)
+    pbj.implicit_solvent.pb_formulation.formulations.direct_stern.lhs(self)
 
 
 def rhs(self):
-    pbj.electrostatics.pb_formulation.formulations.direct_stern.rhs(self)
+    pbj.implicit_solvent.pb_formulation.formulations.direct_stern.rhs(self)
 
 
 def block_diagonal_preconditioner(self):
-    pbj.electrostatics.pb_formulation.formulations.direct_stern.block_diagonal_preconditioner(
+    pbj.implicit_solvent.pb_formulation.formulations.direct_stern.block_diagonal_preconditioner(
         self
     )
 
 
 def mass_matrix_preconditioner(self):
-    pbj.electrostatics.pb_formulation.formulations.direct_stern.mass_matrix_preconditioner(
+    pbj.implicit_solvent.pb_formulation.formulations.direct_stern.mass_matrix_preconditioner(
         self
     )
 
@@ -65,7 +65,7 @@ def calculate_potential(self, rerun_all, rerun_rhs):
     self.ep_stern = ep_stern
 
     if self.stern_object is None:
-        pbj.electrostatics.pb_formulation.formulations.direct_stern.create_stern_mesh(
+        pbj.implicit_solvent.pb_formulation.formulations.direct_stern.create_stern_mesh(
             self
         )
 
@@ -98,7 +98,7 @@ def calculate_potential(self, rerun_all, rerun_rhs):
             self.e_hat_stern = self.ep_stern / self.ep_ex
 
         else:
-            pbj.electrostatics.pb_formulation.formulations.slic.create_ehat_diel(
+            pbj.implicit_solvent.pb_formulation.formulations.slic.create_ehat_diel(
                 self, sigma
             )
             create_ehat_stern(self)
@@ -106,7 +106,7 @@ def calculate_potential(self, rerun_all, rerun_rhs):
 
         calculate_potential_slic(self)
 
-        sigma = pbj.electrostatics.pb_formulation.formulations.slic.solve_sigma(self)
+        sigma = pbj.implicit_solvent.pb_formulation.formulations.slic.solve_sigma(self)
 
         if it != 0:
             phi_L2error = np.sqrt(
